refactor: Extract game start logic into useGameStart hook
