+++ client/src/components/lobby/StartButton.tsx
@@ -1,21 +1,31 @@
 import { Button } from '@/components/ui/Button';
-import { useGameStart } from '@/hooks/useStartButton';
+import { useGameStart } from '@/hooks/game/useGameStart';
+import { useShortcuts } from '@/hooks/useShortcuts';
 import { cn } from '@/utils/cn';
 
 export const StartButton = () => {
-  const { isHost, buttonConfig, handleStartGame, isStarting } = useGameStart();
+  const { startGame, checkCanStart, getStartButtonStatus, isStarting } = useGameStart();
+  const { disabled, title, content } = getStartButtonStatus();
+
+  useShortcuts([
+    {
+      key: 'GAME_START',
+      action: () => startGame(),
+    },
+  ]);
+
   return (
     <Button
-      onClick={handleStartGame}
-      disabled={buttonConfig.disabled || isStarting}
-      title={buttonConfig.title}
+      onClick={startGame}
+      disabled={disabled || isStarting}
+      title={title}
       className={cn(
         'h-full rounded-none border-0 text-xl',
         'sm:rounded-2xl sm:border-2 lg:text-2xl',
-        !isHost && 'cursor-not-allowed opacity-50 hover:bg-violet-500',
+        !checkCanStart() && 'cursor-not-allowed opacity-50 hover:bg-violet-500',
       )}
     >
-      {isStarting ? '곧 게임이 시작됩니다!' : buttonConfig.content}
+      {isStarting ? '곧 게임이 시작됩니다!' : content}
     </Button>
   );
 };
